requirements.txt
browser-use>=1.1.0
langchain-openai>=0.0.3
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0  # For DeepSeek API calls
